app/requirements.txt
pytesseract==0.3.13
requests==2.32.3
Pillow==10.4.0
uvicorn==0.30.6
fastapi==0.95.0
numpy==1.24.1
uzbek-latin-cyrillic-converter==0.2
pandas==2.0.0
pycld2==0.41
celery==5.2.7
kombu==5.2.4
PyJWT==2.7.0
flower==1.2.0
pdf2image==1.17.0
PyPDF2==3.0.1
python-multipart
Spire.Doc==12.7.1
PyMuPDF==1.24.10
python-pptx==1.0.2
Aspose.Slides==24.9.0
tika==2.6.0
beautifulsoup4==4.12.3
tika-client==0.9.0
chardet==5.2.0